+++ backend/src/ban/repository/ban.repository.ts
@@ -3,9 +3,9 @@ import Lent from 'src/entities/lent.entity';
 import { Repository } from 'typeorm';
 import BanLog from '../../entities/ban.log.entity';
 import { IBanRepository } from './ban.repository.interface';
-import { Transactional } from 'src/decorator/transactional.decorator';
-import { Propagation } from 'src/enums/propagation.enum';
-import { IsolationLevel } from 'src/enums/isolation.enum';
+import { Transactional } from 'src/transaction/src/transactional.decorator';
+import { Propagation } from 'src/transaction/src/propagation.enum';
+import { IsolationLevel } from 'src/transaction/src/isolation.enum';
 
 export class BanRepository implements IBanRepository {
   constructor(
