+++ backend/src/blackhole/blackhole.service.ts
@@ -15,10 +15,10 @@ import { UserDto } from 'src/dto/user.dto';
 import { UserService } from 'src/user/user.service';
 import { LentService } from 'src/lent/lent.service';
 import { BlackholeTools } from './blackhole.component';
-import { Transactional } from 'src/decorator/transactional.decorator';
-import { Propagation } from 'src/enums/propagation.enum';
-import { IsolationLevel } from 'src/enums/isolation.enum';
-import { runOnTransactionComplete } from 'src/transaction/hooks.transaction';
+import { Transactional } from 'src/transaction/src/transactional.decorator';
+import { Propagation } from 'src/transaction/src/propagation.enum';
+import { IsolationLevel } from 'src/transaction/src/isolation.enum';
+import { runOnTransactionComplete } from 'src/transaction/src/hooks.transaction';
 import LentType from 'src/enums/lent.type.enum';
 import { CabinetInfoService } from 'src/cabinet/cabinet.info.service';
 import CabinetStatusType from 'src/enums/cabinet.status.type.enum';
