+++ backend/src/lent/lent.component.ts
@@ -6,10 +6,10 @@ import { BanService } from '../ban/ban.service';
 import { CabinetInfoService } from '../cabinet/cabinet.info.service';
 import { LentService } from './lent.service';
 import { ILentRepository } from './repository/lent.repository.interface';
-import { Transactional } from 'src/decorator/transactional.decorator';
-import { Propagation } from 'src/enums/propagation.enum';
-import { IsolationLevel } from 'src/enums/isolation.enum';
-import { runOnTransactionComplete } from 'src/transaction/hooks.transaction';
+import { Transactional } from 'src/transaction/src/transactional.decorator';
+import { Propagation } from 'src/transaction/src/propagation.enum';
+import { IsolationLevel } from 'src/transaction/src/isolation.enum';
+import { runOnTransactionComplete } from 'src/transaction/src/hooks.transaction';
 import { UserDto } from 'src/dto/user.dto';
 import LentExceptionType from 'src/enums/lent.exception.enum';
 import { ConfigService } from '@nestjs/config';
