+++ backend/src/transaction/src/index.ts
@@ -1,0 +1,20 @@
+export {
+  initializeTransactionalContext,
+  addTransactionalDataSource,
+  getDataSourceByName,
+  deleteDataSourceByName,
+} from './common.transaction';
+export {
+  runOnTransactionCommit,
+  runOnTransactionRollback,
+  runOnTransactionComplete,
+} from './hooks.transaction';
+export { Transactional } from './transactional.decorator';
+export { Propagation } from './propagation.enum';
+export { IsolationLevel } from './isolation.enum';
+export { runInTransaction } from './run.in.transaction';
+export {
+  wrapInTransaction,
+  WrapInTransactionOptions,
+} from './wrap.in.transaction';
+export { TransactionalError } from './errors.transaction';
