+++ backend/src/transaction/src/propagation.enum.ts
