+++ backend/src/transaction/src/run.in.transaction.ts
@@ -1,0 +1,13 @@
+import {
+  WrapInTransactionOptions,
+  wrapInTransaction,
+} from './wrap.in.transaction';
+
+export const runInTransaction = <
+  Func extends (this: unknown) => ReturnType<Func>,
+>(
+  fn: Func,
+  options?: WrapInTransactionOptions,
+) => {
+  return wrapInTransaction(fn, options)();
+};
